fix(backend): move process event handlers to main entry point
- Moved SIGINT, SIGTERM, uncaughtException, and unhandledRejection handlers from createServer() to src/index.ts
- Prevents duplicate handler registration when createServer() is called multiple times (e.g., in tests)
- Added server null checks in handlers to handle cases where server creation fails
- Removed unused disconnectDatabase import from server.ts
